\name{LdaFuncs-package}
\alias{LdaFuncs-package}
\alias{LdaFuncs}
\docType{package}
\title{
    LdaFuncs
}
\description{
    An Implementation of the CVB0 algorithm for inference on Latent Dirichlet Allocation topic models.
}
\details{
    Implements weighting for each unique term, allowing words to differ in terms of how much influence they have in both topic formation and document-topic assignment.
}
\author{
Nathan Wycoff

Maintainer: Nathan Wycoff <nathanbrwycoff@gmail.com>
}
\references{
    Blei D.M., Ng A.Y., Jordan M.I. (2003).
    Latent Dirichlet Allocation.

    Arthur Asuncion, Max Welling, Padhraic Smyth, Yee Whye Teh
    On Smoothing and Inference for Topic Models
}
\keyword{TopicModeling}
\seealso{
    tm, TopicModel
}
\examples{
  \dontrun{
     ## Optional simple examples of the most important functions
     ## These can be in \dontrun{} and \donttest{} blocks.   
  }
}
