% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wLDA.R
\name{wLDA}
\alias{wLDA}
\title{Do inference on a Latent Dirichlet Allocation model using Collapsed Variational Bayes with unequal token weighting. This code is optimized for speed over stability.}
\usage{
wLDA(docs, alpha, eta, K, V, thresh = 1e-06, iters = 100, seed, weights)
}
\arguments{
\item{docs}{A list of integer vectors, one integer vector for each document. Each element in a integer vector represents a word and indicates the index of each token in the vocabulary. This index should begin at 1.}

\item{alpha}{The hyperparameter for the Dirichlet prior associated with the document by term matrix. May be a vector of length K. If a single value is passed, a K-vector will be filled with that value.}

\item{eta}{The hyperparameter for the Dirichlet prior associated with the document by topic matrix. May be a vector of length V. If a single value is passed, a V-vector will be filled with that value.}

\item{K}{The number of topics, an integer >= 2}

\item{V}{The size of the vocabuary, an integer. The vocabulary itself does not need to be passed.}

\item{thresh}{The threshold for stopping the iterative algorithm. If the absolute change from one iteration to the next of any element of the word-topic matrices.}

\item{iters}{Maximum iterations. An iteration is defined as one pass over the corpus.}

\item{seed}{Seed for RNG.}

\item{weights}{A vector of positive weights of length V, defining the importance of each word in terms of Topic Formation.}
}
\value{
A list with an element BETA, the topic-by-word matrix, and GAMMA, the document by topic matrix.
}
\description{
Do inference on a Latent Dirichlet Allocation model using Collapsed Variational Bayes with unequal token weighting. This code is optimized for speed over stability.
}
\examples{
#Generate some documents
K <- 2
V <- 4
M <- 20
N.mu <- 10
eta <- rep(1, V)
alpha <- rep(1, K)

ret <- gen.lda(K, V, M, N.mu, eta, alpha)

docs <- ret$docs
Ns <- ret$Ns
BETA <- ret$BETA
THETA <- ret$THETA

#Do inference, see if we got the BETA right
ret <- wLDA(docs, alpha, eta, K, V)

cat('True BETA:')
print(BETA)

cat('Our guess:')
print(ret$BETA)
}
